requirements/base.txt
celery==5.3.4
django==4.2.5
django-celery-beat==2.5.0
django-cloudinary-storage==0.3.0
django-environ==0.11.2
django-jazzmin==2.6.0
django-redis==5.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
gunicorn==21.2.0
mnemonic==0.20
orjson==3.9.7
web3==6.10.0